+++ src/main/java/prometheus/PrometheusException.java
@@ -1,3 +1,5 @@
+package prometheus;
+
 public class PrometheusException extends Exception {
     public PrometheusException(String message) {
         super(message);
