+++ src/main/java/prometheus/command/ExitCommand.java
@@ -1,3 +1,8 @@
+package prometheus.command;
+import prometheus.Storage;
+import prometheus.Ui;
+import prometheus.TaskList;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
