+++ src/main/java/prometheus/task/Event.java
@@ -1,7 +1,9 @@
+package prometheus.task;
+
+import prometheus.PrometheusException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
-import java.time.Duration;
 
 public class Event extends Task {
     protected LocalDateTime from;
