+++ src/main/java/prometheus/task/Todo.java
@@ -1,4 +1,6 @@
-public class Todo extends Task{
+package prometheus.task;
+
+public class Todo extends Task {
     public Todo(String description) {
         super(description);
     }
